Destroy hosts from the web front-end
This adds a new `destroy_host` method to the Destroyer, which uses
the `run_cmd` approach to capture the destroy call output. It also
adds an extra host status for use while the destroy is happening.

There's a little bit more error handling for the Celery tasks, so
exceptions get reported in the deploy log, but this will need to be
more robust.

The 'delete' button now does a destroy, but leaves the host in
'defining' mode. Clicking delete again will remove it from the DB.