Start changes to move naming to its own module
The goal is to abstract the naming of hosts and resources into a separate module,
so that there is a single point of reference when a name or URL is required throughout
the code. A host now has two relevant fields: base_name (specified through the form,
eg "cucumber_cats") and dns_name (including anything that should be added to "ensure"
that the final URL is unique, eg "ucl_cucumber_cats_ravenous_chupacabra_42"). The
base name will be passed to the naming module to provide an appropriate dns name, and
both will be stored in the DB. The dns name can then be used to generate names for
additional resources or URLs, and the base name remains available to eg display
a user-friendly *.cloudlabs.ucl.ac.uk URL. Currently, the dns name just contains
the "ucl" prefix, so randomness still has to be added.