Clarify that Azure 'stopped' status is not yet off
A stopped (but not deallocated) VM still incurs charges, whereas a
dellocated one does not. It would therefore be misleading to show
a stopped machine as actually stopped, especially since the status
could then move back to 'stopping' (while deallocating). I am not
sure whether deallocating happens as a second step or is a wholly
separate operation, but in either case I think it should be fine
to treat Azure 'stopped' as a temporary status, i.e. map it to our
'stopping' Host status.