Update the DB with each VM's live status
Added a periodic task that queries Azure and updates the existing
hosts' statuses in the DB. We will need to run `celery beat` for
this, in addition to starting the Celery workers. At the moment,
the dashboard displays the live status but the buttons depend on
the DB status (which will be updated eventually, anyway). Ideally,
we want some logic to combine the live status with info from the
application, particularly to identify hosts being deployed (which
we cannot do with information from Azure only).