Use new Azure tools class
This makes some of the deployer methods (except for deploy and
destroy) trivial. They could probably be removed and we could
then just have the start, stop etc methods called form the celery
tasks. I am leaving them in for now because we may still want to
do something for which we need the deployer (eg update the DB).
This commit also makes the start/restart methods more consistent
(restart is only for hosts that are already running, start is for
either stopped or running).