Configure the new loggers
Note that we need to configure our own loggers (cloudlabs.*)
specifically, and not the root! This is because Celery overrides
the root logger configuration. It should be possible to turn that
off, but it seems it's not that simple. Setting the congiruation
for our family of loggers seems to work, however, and we only need
to set the "parent" logger (cloudlabs) rather than each logger
(e.g. cloudlabs.admin) explicitly.