Configure logger from YAML file
This also makes some minor changes to the log format and settings.
NOTE: It's possible that the location of the log config file has
to be given in a different way, in case Flask is started from a
different location (eg through Apache). Will need to check this
on the servers!