Adjust logging to work on servers
Configuring logging is a little tricky, because on the servers
(both staging and production) the app is executed as a different
user, who does not have permissions on the folder where the code
is. We therefore have to place the log file somewhere where that
user will definitely have permissions, e.g. their home directory
(/var/log is also tricky, because the app is not run with sudo),
at least not when setting up the database (perhaps when serving?).
This commit doesn't change the location of the log file when the
app is run under the development configuration, which should
include our local development setup. In these cases, the log will
still be written in the directory of the code.