Expose Azure URL again for tests
When testing on Travis or a local machine, the URL mapping will
fail as the request will not be allowed through the firewall. The
failure in itself is not a problem (it just adds a long entry), but
the tests also "ping" the machine using the "final" (UCL) URL, at
which point they will fail. This change adds a new property that
replicates the old URL (before the DNS mapping was implemented),
and should be used in tests or for debugging.