Decouple task-checking from status-checking
This is to make sure that the correct deletion action (whether hard
or no delete at all) is taken, regardless of whether a task is
still recorded in the database. This shouldn't happen, but in some
cases previous hosts were showing up as still having an associated
task, and therefore could not be deleted correctly.